+++ docs/conf.py
@@ -48,6 +48,7 @@ else:
         "sphinx.ext.autodoc",
         "sphinx_autodoc_typehints",
         "sphinx-jsonschema",
+        "sphinxcontrib.mermaid",
     ]
 
     # meta
@@ -104,7 +105,7 @@ else:
                 return
             rst = pypandoc.convert_text(description, "rst", format="md")
             if isinstance(container, list):
-                container.append(self._linme(self._cell(rst)))
+                container.append(self._line(self._cell(rst)))
             else:
                 self.state.nested_parse(
                     self._convert_content(rst), self.lineno, container